example/cli/pyproject.toml
[tool.poetry]
name = "honcho-cli-example"
version = "0.1.0"
description = "CLI example for honcho"
authors = ["Vineeth Voruganti <13438633+VVoruganti@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
honcho-ai = {path = "../../sdk", develop=true}
langchain-openai = "^0.0.5"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
langchain = "^0.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
